feat(passengers): implement custom hook for managing passenger forms
