+++ src/app/(frontend)/flights/booking/passengers/page.client.tsx
@@ -9,12 +9,10 @@ import {
   type ContactInfoValidationErrors,
   validateContactInfo,
 } from "@/components/flights/contact-info-card";
-import {
-  PassengerFormCard,
-  type PassengerFormData,
-} from "@/components/flights/passenger-form-card";
+import { PassengerFormCard } from "@/components/flights/passenger-form-card";
 import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
+import { usePassengerForms } from "@/hooks/use-passenger-forms";
 import { createOrderAction } from "@/lib/actions/orders";
 
 import type { SavedPassenger } from "./queries";
@@ -33,21 +31,17 @@ export function BookingPassengersPageClient({
   savedPassengers,
 }: BookingPassengersPageClientProps) {
   const router = useRouter();
-  const [selectedPassengers, setSelectedPassengers] = useState<string[]>([]);
-  // Map to track which passenger form is filled by which saved passenger
-  const [passengerFormMapping, setPassengerFormMapping] = useState<
-    Map<number, string>
-  >(new Map());
-  const [passengers, setPassengers] = useState<PassengerFormData[]>([
-    {
-      chineseName: "",
-      englishFirstName: "",
-      englishLastName: "",
-      documentType: "id_card",
-      documentNumber: "",
-      phone: "",
-    },
-  ]);
+
+  // Use custom hook for passenger form management
+  const {
+    passengers,
+    selectedPassengers,
+    handleAddPassenger,
+    handleRemovePassenger,
+    handlePassengerChange,
+    handleSelectSavedPassenger,
+  } = usePassengerForms(savedPassengers);
+
   const [contactInfo, setContactInfo] = useState<ContactInfo>({
     method: "email",
     email: "",
@@ -57,151 +51,6 @@ export function BookingPassengersPageClient({
     useState<ContactInfoValidationErrors>({});
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [submitError, setSubmitError] = useState<string | null>(null);
-
-  const handleAddPassenger = () => {
-    setPassengers([
-      ...passengers,
-      {
-        chineseName: "",
-        englishFirstName: "",
-        englishLastName: "",
-        documentType: "id_card",
-        documentNumber: "",
-        phone: "",
-      },
-    ]);
-  };
-
-  const handleRemovePassenger = (index: number) => {
-    if (passengers.length > 1) {
-      // Check if this form was filled by a saved passenger
-      const savedPassengerId = passengerFormMapping.get(index);
-      if (savedPassengerId) {
-        // Deselect the saved passenger
-        setSelectedPassengers(
-          selectedPassengers.filter(id => id !== savedPassengerId)
-        );
-      }
-
-      // Remove the passenger form
-      setPassengers(passengers.filter((_, i) => i !== index));
-
-      // Update mapping: remove the deleted index and adjust indices after it
-      const newMapping = new Map<number, string>();
-      passengerFormMapping.forEach((passengerId, formIndex) => {
-        if (formIndex < index) {
-          // Keep indices before the deleted one
-          newMapping.set(formIndex, passengerId);
-        } else if (formIndex > index) {
-          // Shift down indices after the deleted one
-          newMapping.set(formIndex - 1, passengerId);
-        }
-        // Skip the deleted index
-      });
-      setPassengerFormMapping(newMapping);
-    }
-  };
-
-  const handlePassengerChange = (
-    index: number,
-    field: keyof PassengerFormData,
-    value: string
-  ) => {
-    const updated = [...passengers];
-    updated[index][field] = value;
-    setPassengers(updated);
-
-    // If this form was filled by a saved passenger, unlink it
-    const savedPassengerId = passengerFormMapping.get(index);
-    if (savedPassengerId) {
-      // Remove from selected passengers
-      setSelectedPassengers(
-        selectedPassengers.filter(id => id !== savedPassengerId)
-      );
-      // Remove from mapping
-      const newMapping = new Map(passengerFormMapping);
-      newMapping.delete(index);
-      setPassengerFormMapping(newMapping);
-    }
-  };
-
-  const handleSelectSavedPassenger = (passengerId: string) => {
-    const isCurrentlySelected = selectedPassengers.includes(passengerId);
-
-    if (isCurrentlySelected) {
-      // Deselect: Remove from selected list and clear the form
-      setSelectedPassengers(
-        selectedPassengers.filter(id => id !== passengerId)
-      );
-
-      // Find which form index was filled by this saved passenger
-      const formIndex = Array.from(passengerFormMapping.entries()).find(
-        ([, id]) => id === passengerId
-      )?.[0];
-
-      if (formIndex !== undefined) {
-        // Clear the form
-        const updated = [...passengers];
-        updated[formIndex] = {
-          chineseName: "",
-          englishFirstName: "",
-          englishLastName: "",
-          documentType: "id_card",
-          documentNumber: "",
-          phone: "",
-        };
-        setPassengers(updated);
-
-        // Remove from mapping
-        const newMapping = new Map(passengerFormMapping);
-        newMapping.delete(formIndex);
-        setPassengerFormMapping(newMapping);
-      }
-    } else {
-      // Select: Add to selected list and auto-fill a form
-      setSelectedPassengers([...selectedPassengers, passengerId]);
-
-      // Find the saved passenger data
-      const savedPassenger = savedPassengers.find(p => p.id === passengerId);
-      if (!savedPassenger) return;
-
-      // Find the first empty form or create a new one
-      let targetFormIndex = passengers.findIndex(
-        p =>
-          !p.chineseName &&
-          !p.englishFirstName &&
-          !p.englishLastName &&
-          !p.documentNumber
-      );
-
-      const newPassengerData: PassengerFormData = {
-        chineseName: savedPassenger.chineseName || "",
-        englishFirstName: savedPassenger.englishFirstName || "",
-        englishLastName: savedPassenger.englishLastName || "",
-        documentType: savedPassenger.documentType,
-        documentNumber: savedPassenger.documentNumber,
-        phone: savedPassenger.phone || "",
-      };
-
-      let updatedPassengers: PassengerFormData[];
-      if (targetFormIndex === -1) {
-        // No empty form found, add a new one with the data
-        targetFormIndex = passengers.length;
-        updatedPassengers = [...passengers, newPassengerData];
-      } else {
-        // Fill the existing empty form
-        updatedPassengers = [...passengers];
-        updatedPassengers[targetFormIndex] = newPassengerData;
-      }
-
-      setPassengers(updatedPassengers);
-
-      // Update mapping
-      const newMapping = new Map(passengerFormMapping);
-      newMapping.set(targetFormIndex, passengerId);
-      setPassengerFormMapping(newMapping);
-    }
-  };
 
   const handleNext = async () => {
     // Clear previous errors
